pyproject.toml
[build-system]
requires = ["setuptools>=62.0", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "dingo-gw"
authors = [
    {name = "Maximilian Dax", email = "maximilian.dax@tuebingen.mpg.de"},
    {name = "Stephen Green", email = "Stephen.Green2@nottingham.ac.uk"},
]
description = "Deep inference for gravitational-wave observations"
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10,<3.12"
dynamic = ["version"]

classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent"
]

dependencies = [
    "astropy",
    "bilby",
    "bilby_pipe",
    "configargparse",
    "corner",
    "glasflow",
    "gwpy",
    "h5py",
    "lalsuite>=7.15",
    "matplotlib",
    "nptyping",
    "numpy",
    "pandas",
    "pycbc",
    "pycondor",
    "pyyaml",
    "requests",
    "rich",
    "scikit-learn",
    "scipy",
    "threadpoolctl",
    "torch",
    "torchdiffeq",
    "torchvision",
    "tqdm",
]

[tool.setuptools.packages.find]
include = ["dingo*"]
namespaces = false

[tool.setuptools_scm]
write_to = "dingo/_version.py"

[tool.pytest.ini_options]
markers = [
    "slow: marks tests as slow",
    "asimov: marks tests for Asimov/LIGO pipeline integration",
]
addopts = "-m 'not asimov'"

[project.urls]
homepage = "https://github.com/dingo-gw/dingo"

[project.optional-dependencies]
dev = [
    "black",
    "ipykernel",
    "linkify-it-py",
    "myst-nb",
    "myst-parser",
    "pandas-stubs>=2.3.3.251219",
    "pytest",
    "scipy-stubs>=1.15.3.0",
    "sphinx",
    "sphinx-math-dollar",
    "sphinx-rtd-theme",
    "sphinxcontrib-mermaid",
    "sphinxcontrib-bibtex",
]
wandb = ["wandb"]
pyseobnr = ["pyseobnr"]

[dependency-groups]
dev = [
    "black",
    "ipykernel",
    "pytest",
]

docs = [
    "linkify-it-py",
    "myst-nb",
    "myst-parser",
    "sphinx",
    "sphinx-math-dollar",
    "sphinx-rtd-theme",
    "sphinxcontrib-mermaid",
    "sphinxcontrib-bibtex",
]

typing = [
    "pandas-stubs>=2.3.3.251219",
    "scipy-stubs>=1.15.3.0",
]

[project.scripts]
dingo_append_training_stage = "dingo.gw.training:append_stage"
dingo_build_svd = "dingo.gw.dataset.utils:build_svd_cli"
dingo_estimate_psds = "dingo.gw.noise.asd_estimation:download_and_estimate_cli"
dingo_evaluate_multibanded_domain = "dingo.gw.dataset.evaluate_multibanded_domain:main"
dingo_generate_asd_dataset = "dingo.gw.noise.generate_dataset:generate_dataset"
dingo_generate_dataset = "dingo.gw.dataset.generate_dataset:main"
dingo_generate_dataset_dag = "dingo.gw.dataset.generate_dataset_dag:main"
dingo_generate_synthetic_asd_dataset = "dingo.gw.noise.synthetic.generate_dataset:main"
dingo_ls = "dingo.gw.ls_cli:ls"
dingo_merge_datasets = "dingo.gw.dataset.utils:merge_datasets_cli"
dingo_merge_asd_datasets = "dingo.gw.noise.utils:merge_datasets_cli"
dingo_pipe = "dingo.pipe.main:main"
dingo_pipe_importance_sampling = "dingo.pipe.importance_sampling:main"
dingo_pipe_generation = "dingo.pipe.data_generation:main"
dingo_pipe_plot = "dingo.pipe.plot:main"
dingo_pipe_pp_test = "dingo.pipe.pp_test:main"
dingo_pipe_sampling = "dingo.pipe.sampling:main"
dingo_pt_to_hdf5 = "dingo.core.utils.pt_to_hdf5:main"
dingo_result = "dingo.pipe.dingo_result:main"
dingo_train = "dingo.gw.training:train_local"
dingo_train_condor = "dingo.gw.training.train_pipeline_condor:train_condor"

[project.entry-points."asimov.pipelines"]
dingo = "dingo.asimov.asimov:Dingo"
